+++ tests/cli.rs
@@ -12,6 +12,7 @@ use gherkin::Step;
 mod common;
 
 use common::TEMPDIR_PLACEHOLDER;
+use common::assert_avro_codec;
 use common::assert_output_contains;
 use common::assert_valid_parquet_file;
 use common::get_row_count;
@@ -370,6 +371,12 @@ fn that_file_should_be_valid_parquet(world: &mut CliWorld) {
         .as_ref()
         .expect("No file has been set; use 'the file \"...\" should exist' first");
     assert_valid_parquet_file(path);
+}
+
+#[then(regex = r#"^the file "(.+)" should be a valid Avro file with codec "(.+)"$"#)]
+fn file_should_be_avro_with_codec(world: &mut CliWorld, path: String, expected_codec: String) {
+    let path_resolved = resolve_path(world, &path);
+    assert_avro_codec(&path_resolved, &expected_codec);
 }
 
 #[then(regex = r#"^the file "(.+)" should contain:$"#)]